+++ src/main/java/frc/trigon/robot/subsystems/armelevator/ArmElevatorConstants.java
@@ -1,4 +1,4 @@
-package frc.trigon.robot.subsystems.arm;
+package frc.trigon.robot.subsystems.armelevator;
 
 import com.ctre.phoenix6.configs.CANcoderConfiguration;
 import com.ctre.phoenix6.configs.TalonFXConfiguration;
@@ -334,7 +334,7 @@ public class ArmElevatorConstants {
         REST(Rotation2d.fromDegrees(0), 0.603, null, false, 0.7),
         REST_WITH_CORAL(Rotation2d.fromDegrees(180), 0.603, null, false, 0.7),
         REST_WITH_ALGAE(Rotation2d.fromDegrees(90), 0.603, null, false, 0.7),
-        REST_FOR_CLIMB(Rotation2d.fromDegrees(0), 0.603, null, false, 0.7),
+        REST_FOR_CLIMB(Rotation2d.fromDegrees(180), 0.603, null, false, 0.7),
         LOAD_CORAL(Rotation2d.fromDegrees(0), 0.5519, REST, true, 0.7),
         EJECT(Rotation2d.fromDegrees(60), 0.603, null, false, 0.7),
         SCORE_L1(Rotation2d.fromDegrees(70), 0.4, null, false, 1),
